argument/property: Handle all array types in set_property
For a very long time, we have only handled GStrv correctly.
This means that trying to set a GArray/GPtrArray/GByteArray
property would just fail. In order to handle this correctly
we need to use the introspection data to see what *type* of
array we are dealing with.

XXX: Note that this is incomplete - C arrays in particular are not
handled correctly because we have to throw away the size information,
meaning that freeing them on the error path leaks since there's
no safe way to free the elements.