gi: Factor out pygi_set_property_gvalue_from_property_info
Also harmoinze the checks in pygi_set_property_value with
those from the old PyGObject code.

pygi_set_property_value will first try to set properties using
the introspection data (if the GObject had some GITypeInfo
associated with it), otherwise if the object was dynamically
created, we fall back to being able to set a more limited
set of properties.