gi: Use pygi_set_gvalue_for_pspec to set properties on construction
For some reason, setting properties on construction used the
old non-introspection codepath, meaning that we would do very strange
things when setting properties of arrays, for instance, by
setting an array of GValue, which would later crash when we try
to read it because that's obviously not what the underlying
object expected.

This commit makes the construction path use pygi_set_gvalue_for_pspec,
which will first try to use the introspection data to set the
property and then fall back to the non-introspection path
when that isn't available.