property: Use g_value_take_* for container types
This is a bit tricky. For the other types, whether or not
the GIArgument owns the underlying value depends on the
transfer argument, though in the case of setting properties
we have to create a new container from passed in Python value
which we then own a reference on.

g_value_set_boxed will take ownership of the boxed type, meaning
that we have to get rid of our own ownership somehow - the best way
to do this is to transfer the ref to the GValue, which will then
get rid of the ref once it goes away. It is up to the underlying
object to take a reference here.

XXX: That said, for properties that are transfer full, its not
totally clear what happens here.