argument: Set the free_func of GPtrArray
Add some helpers to set free_funcs of various container types

The usual pattern for most container types is to just
g_*_unref them once the owner is done with them. For some
containers, it is possible to accept a "free func" which
destroys the elements once the container's ref drops to
zero. This function returns some common free-functions for
various element types (but not for all of them, in particular
it isn't possible to set a free func for element types
where the freeing function requires some additional information
in a closure).