argument: Set the key_free_func and value_free_func of GHashTable
This way we don't have to manually traverse the hashtable on
release to free everything. On the upside this means that
we can then pass the hashtable directly to some object
and it can call g_hash_table_unref on it without worrying about
the contents. On the downside, there are a few less types that
we can handle in the error paths that call _pygi_argument_release