Fix regression in marshalling partial() objects
In a4880dbc4575fadc0e3 a special case for partial() was added to handle gtk4 template
callbacks. This in turn broken normal usage of partial objects.

To work around that add a special marker in the gtk template code for now until
we find a better fix.

Also adds a test so this doesn't happen again.

Fixes #464