Expose GObject.Object.run_dispose()
Up until now this raised an exception "This method is currently unsupported.".

With Gtk.Widget.destroy() gone in gtk4 and that method often being used to remove
references to other objects and breaking cycles this is the next best thing
on a lower level and should make porting a bit easier.

Fixes #470