overrides: Add EventLoop integration points into overrides
This adds EventLoop integration in order to mark the event loop as
running when a main context iterating API is called. Note that for the
simple APIs to only do one iteration the EventLoop is paused (i.e. it
will not dispatch). No one should do this, but it might happen when e.g.
porting and this should create a well-defined behaviour.