docs: add a page about system dependencies and their versions over time
As pointed out in https://gitlab.gnome.org/GNOME/pygobject/-/issues/590 it
can happen that pygobject can't be installed because system dependencies are
too old, and pip can't help there. The only solution is to pin pygobject to an
older version in that case.

To help users decide which version they should target and to give some info on
which dependencies are needed add a small page to the docs.

Fixes #590