fundamental: Implement weak references more properly
It appears that PyObject_ClearWeakRefs may only be called on PyPy if
there are actually weak references to the object. Note that to correctly
track that, one also needs a separate weaklist to store them.

Add a weaklist entry, setup the tp_weaklistoffset and only call
PyObject_ClearWeakRefs if there are weak references. Also add a test
that actually uses weak references to ensure it doesn't crash either.