CI: use pypy from the Ubuntu PPA instead of pyenv
As pointed out in !299 the libffi shipped with the portable pypy version
conflicts with the system libffi, and could lead to crashes.

Instead of the pyenv version just use the official PPA which contains
a pypy linked against system libs. That has the downside that the version
is no longer pinned and docker image rebuilds will lead to newer pypy versions
being included.