CI: do everything in a venv and with the same pip version
After !301 the pypy version has a different pip version compared to the
oder pyenv cpython installations and also its pip installs to the system
which comes with new restrictions with newer pip (requires --break-system-packages).

To make the test environment for all interepreter versions the same create
and active a venv first and also update pip in there.

This is motivated by the pip in 22.04 not supporting "--config-settings",
which is required in !302