async: do not depend on asyncio to create a usable Future
This change the gi.Async type to be usable without an EventLoop. Main
thing is to not go through loop.call_soon and instead use the correct
main context directly for the callback if needed.

With this, we also avoid doing an extra mainloop iteration and call the
callbacks directly. This is reasonable as GTask already guarantees that
the callback is delayed using an idle handler if it would otherwise be
delivered from the same maincontext iteration that it was created on.