async: Correctly normalize the exception
The exception value returned by PyErr_Fetch may not be an exception
instance but rather the parameter for the exception class __init__
function. The correct way to ensure it is an instance is to call
PyErr_NormalizeException.

Note that in our case, this call will always be a no-op as the called
method is a function call into C which can only generate a GError and
it always has a properly instantiated value already.

Fixes: b891e404b958 ("async: Add a new async type that is an awaitable for a _finish call")
Closes: #641