overrides: do not duck-type GLibEventLoop and policy
Only return the current event loop if it really has the expected type.
Before, we were simply returning it if it has the correct _context
attribute, but this is not sufficient as other implementations (in
particular GBulb) had the GMainContext stored under the same attribute
name.

Be more strict about the types including verifying the policy itself to
not implicitly create a new EventLoop (e.g. inside a helper thread).

Fixes: #644