overrides: Add a convenience API for task creation
It can be dangerous to just create a task without holding a reference to
it. Also, it generally makes sense to not stop an application before all
tasks have completed.

Add a convenience API to Gio.Application to create a task for a
coroutine. All it does is keeping a reference to the task and ensuring
that the application does not quit before the task is done.

Providing the API avoids two pitfalls with asyncio Task handling. It
also gives us the opportunity to inform people that they may need to
cancel tasks and the pitfalls that exist there.