events: Support setting idle priority for tasks
The asyncio call_soon handler is pretty much identical to g_idle_add
conceptually. However, there is g_idle_add_full to set a priority, which
python does not really have.

As priorities are rather useful, add the ability to set the priority of
a task. Do this by simply trying to read the _glib_idle_priority
attribute of the current task, and, if set, use that for registering a
call_soon handler.

Doing this correctly is a bit more involved as we need a second GSource
to handle the different priority. And, we also need to make sure that
only one of them can be dispatching at a time.