Set Py_TPFLAGS_METHOD_DESCRIPTOR on PyGIFunctionInfo_Type
This indicates that the type has descriptor behaviour similar to the
regular function type. This signals that it is safe to convert a call
like foo.method(args...) to method(foo, args...) if the interpreter
thinks that would be faster.

The descr_get method for PyGIFunctionInfo_Type did somewhat more than
this though, so I've removed that code and wrapped constructors in
classmethod and non-methods in staticmethod.