gi: Add a custom error for attempting to marshal str as an array
This previously failed with a somewhat cryptic error:

    >>> Gio.MemoryOutputStream.new_resizable().write('hello')
    Traceback (most recent call last):
      File "<python-input-2>", line 1, in <module>
        Gio.MemoryOutputStream.new_resizable().write('hello')
        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^
    TypeError: Must be number, not str

A number, huh? This is because _pygi_marshal_from_py_array attempts to
iterate over the sequence object, and strings, when iterated, appear to
consist of nested single-character strings.

Instead, detect this case explicitly, and guide the user towards
converting the string to bytes.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>