gi/overrides/Gio: Use Platform prefix for platform-only symbols
In previous versions of GLib, platform-specific symbols such as
GUnixMountMonitor were mapped in the Gio namespace as Gio.UnixMountMonitor,
while since commit 6a234a92b1b614b7434340c4b15 we create wrappers such as
Gio.MountMonitor.

This is not correct, and does not serve the initial purpose of providing
a backward compatible wrapper.

So, use the same logic that we had before: if the GType of a symbol starts
with G{Unix,Win32} we use the platform specific name as prefix of the
wrapper type, so that it will be Gio.{Unix,Win32}TypeName

See: https://gitlab.gnome.org/GNOME/gjs/-/merge_requests/1027