Use a (temporary) object to make sure do_dispose is called
Make sure we have a Python object on which we can call `do_dispose`.

The temporary object is only created if we know there's a `do_dispose`
method.

For PyPy we do not create a new wrapper object, because that will be
destroyed asynchronously, leaving us with an extra reference on the
GObject.