overrides/Gio: Simplify the wrapper name computation using the info type name
In case the type name is not provided or if it's starting with the
platform name value, we should use it.

By doing this we can drop some custom code that was done to handle cases
such as UnixInputStreamClass.

This is all already tested and we are still working as expected