overrides/Gio: Warn if trying to reuse a symbol in newer GLib
In case the wrapper name is matching the default name, and we're in a
recent enough version of Gio, we must warn that the wrapper name is
not part of Gio or we are going to skip it.

For version 2.86 we could even assert this, but let's not break users if
in future some symbol is matching

But let's just make it clearer.