overrides/GLib: Add GLib wrappers for GLibUnix symbols
In GLib 2.88 we are dropping the GLib platform-specific symbols from the
main GLib namespace upstream, but again we don't want to break current
applications that may use them, so let's handle this breakage in the
bindings by adding wrappers.

We use the same logic of Gio here.

However, differently from what happens in Gio here we do not have any
GType, and so no introspectable classes, thus we do not have to do any
extra override to support the older versions